+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   code
 - Add padding length check as described by PKCS#1 v1.5 (Fixes
   [#164](https://github.com/sybrenstuvel/python-rsa/issues/164))
+- Reuse of blinding factors to speed up blinding operations.
+  Fixes [#162](https://github.com/sybrenstuvel/python-rsa/issues/162).
 
 
 ## Version 4.4 & 4.6 - released 2020-06-12